regulator: userspace-consumer: allow boot-on with non-exclusive regulator get
If another consumer must be provided for the target regulator, such as
in the case where a workaround (vbus_sel-imx6.c) for the i.MX 6 errata
ERR010661 needs to be added, cannot do an exclusive get of the regulator
(devm_regulator_bulk_get_exclusive).

As described in commit 5c51d4afcf3fd36159713556402e16cfab794ae9, the
reason for devm_regulator_bulk_get_exclusive is to prevent
regulator_init_complete_work from automatically disabling the regulator.

In regulator_init_complete_work, the regulator is automatically disabled
when the use_count of regulator_dev struct is 0.
For example, to set use_count to non-zero, one of the following is
required.
 - Use devm_regulator_bulk_get_exclusive
 - Set regulator-always-on flag
 - Enable the regulator

In situations where the boot-on is needed, enabling the regulator first
will prevent the regulator from being automatically disabled.

For situations where the target regulator must have another consumer and
the regulator needs to be boot-on, this commit provides the following
device tree interfaces.
- Switchable to normal devm_regulator_bulk_get (no-exclusive)
- Switchable to boot-on (init-on and autoswitch)

Change-Id: Ie1cb4afee573c8e36ae72f32f4a631c160969f5b
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>