Revert "x2_defconfig: enable apparmor"
This reverts commit b6c44ce788dd6f6b92c39a2ee93215e565946802.

Enabling apparmor brings a ~7% performance penality, which is sensible
enough for systems such as Armadillo 900.
(This can be brought down to 3% by building apparmor but removing from
the default lsm list, so that it is disabled and can be turned on by
optargs)

There is also no clear and easy way to use it, so the plan will be to
only built it on the -debug kernel which is not meant to be fast, so
people who want to use it can first test it and build their own kernel
for the time being.
Longer term, we will need to support multiple kernels (6.12 kernel,
etc), so we will provide a variant for it.

Change-Id: Ica3e7b3b64a90059558b248f926d157f63999f8c
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>