pwm-imx-tpm: Resolve duty cycle errors in EPWM
In Edge-Aligned PWM (EPWM) mode, when setting the period, subtract 1
from the count corresponding to the period and write the result to the
MOD register. Therefore, when calculating the count corresponding to
duty cycle (= CnV), must multiply the ratio of duty cycle
by the value obtained by adding 1 to MOD.

Prior to this commit, the duty cycle ratio was multiplied by MOD,
resulting in a 1-resolution discrepancy between the specified duty cycle
and the actual duty cycle.
It is noteworthy that when the duty ratio was set to 100%
(period==duty_cycle), the actual PWM waveform did not reach a duty ratio
 of exactly 100% (always High).

Change-Id: I8b19a98f89b2f800887887d739448ec420e0805c