gpio: imx-rpmsg: set irq_type to NONE at resume
Originally, irq_type was not set at resume, so the irq_type used when
calling set_wake differed between the first sleep and subsequent sleeps.

At the first sleep, `irq_type=0` when `set_wake` is called, so it is
guarded, and a message is sent to the M33 side during `set_type`.
Conversely, at the second and subsequent sleeps, `irq_type=3` when
`set_wake` is called, so a message is sent to the M33 side during both
`set_wake` and `set_type`.

While sending multiple messages isn't a major issue, the differing
behavior between the first sleep and subsequent sleeps of this
driver and M33 could introduce unexpected bugs and complicate debugging.

Therefore, to ensure consistent sleep behavior regardless of count, set
the same irq_type (=0) at resume as before suspend.

Change-Id: I0e6108968ec5da9e86fc8879a481badc381d7f8f
Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>