arm64: dts: add armadillo_900-i2c7.dtbo
armadillo_900-i2c7.dtbo is used for a9e sd installer to detect ext
boards such as Wi-SUN ext board. The a9e's ext board has an eeprom
connected via i2c that the a9e can read to detect what type of ext
board is connected.

However, this requires that the i2c be muxed. Since special processing
is required in u-boot to determine whether the board is an a900 or an
a9e, we will create an i2c-only dtbo and apply it without checking.

Change-Id: I9a6a0d1afb06907832ff50789b88bf8fddfe0b92
Signed-off-by: Kojiro Yamada <kojiro.yamada@atmark-techno.com>